refactor(chrome-extension): keep preferences persistence under a single module
