+++ apps/chrome-extension/src-extension/background.ts
@@ -1,8 +1,8 @@
 /* global chrome  */
 import { broadcastMessage } from '@/chrome/utils'
-import { executeStorageMigrations } from '@/features/textAdaptationPreferencesMigrations'
+import { executeStorageMigrations } from '@/features/textAdaptationPreferencesPersistence'
 
-import { STORAGE_KEY_TEXT_ADAPTATION_PREFERENCES } from '@/entities/textAdaptationPreferences'
+import { STORAGE_KEY_TEXT_ADAPTATION_PREFERENCES } from '@/features/textAdaptationPreferencesPersistence'
 
 chrome.runtime.onInstalled.addListener(async () => {
   console.log('readapt installed or updated')
