--- apps/chrome-extension/src/entities/textAdaptationPreferences/config/storage.ts
@@ -1,1 +1,0 @@
-export const STORAGE_KEY_TEXT_ADAPTATION_PREFERENCES = 'textAdaptationPreferences'
