+++ apps/chrome-extension/src/entities/textAdaptationPreferences/index.ts
@@ -2,8 +2,6 @@ export type { TextAdaptationProfile, TextAdaptationPreferences } from './model/T
 
 export { default as TextAdaptationProfilesDropdown } from './ui/TextAdaptationProfilesDropdown.vue'
 
-export { STORAGE_KEY_TEXT_ADAPTATION_PREFERENCES } from './config/storage'
-
 export { textAdaptationProfileFixture } from './model/fixtures/textAdaptationProfileFixture'
 
 export { useTextAdaptationPreferences } from './model/state/useTextAdaptationPreferences'
