--- apps/chrome-extension/src/features/textAdaptationPreferencesMigrations/index.ts
@@ -1,1 +1,0 @@
-export { executeStorageMigrations } from './storage/executeMigrations'
