+++ apps/chrome-extension/src/features/textAdaptationPreferencesPersistence/config/storage.ts
@@ -1,4 +1,4 @@
-import { STORAGE_KEY_TEXT_ADAPTATION_PREFERENCES } from '@/entities/textAdaptationPreferences'
+export const STORAGE_KEY_TEXT_ADAPTATION_PREFERENCES = 'textAdaptationPreferences'
 
 // version 1
 export const STORAGE_KEY_V1 = 'settings'
