+++ apps/chrome-extension/src/features/textAdaptationPreferencesPersistence/index.ts
@@ -1,1 +1,3 @@
-export { initStateFromStorage } from './storage/initStateFromStorage'
+export { initStateFromStorage } from './storage/initStateFromStorage/initStateFromStorage'
+export { executeStorageMigrations } from './storage/migrations/executeStorageMigrations'
+export { STORAGE_KEY_TEXT_ADAPTATION_PREFERENCES } from './config/storage'
